Gaelo2

+++ GaelO2/database/migrations/2020_07_26_211343_create_centers_table.php
@@ -14,7 +14,7 @@ class CreateCentersTable extends Migration
     public function up()
     {
         Schema::create('centers', function (Blueprint $table) {
-            $table->unsignedInteger('code')->primary();
+            $table->unsignedBigInteger('code')->primary();
             $table->string('name')->unique();
             $table->string('country_code', 2)->nullable(false);
             $table->timestamps();
